// frontends/api/src/generated/v1/models/PaginatedLearningResourceList.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MIT Learn API
 * MIT public API
 *
 * The version of the OpenAPI document: 0.0.1 (v1)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime"
import type { LearningResource } from "./LearningResource"
import {
  LearningResourceFromJSON,
  LearningResourceFromJSONTyped,
  LearningResourceToJSON,
  LearningResourceToJSONTyped,
} from "./LearningResource"

/**
 *
 * @export
 * @interface PaginatedLearningResourceList
 */
export interface PaginatedLearningResourceList {
  /**
   *
   * @type {number}
   * @memberof PaginatedLearningResourceList
   */
  count: number
  /**
   *
   * @type {string}
   * @memberof PaginatedLearningResourceList
   */
  next?: string | null
  /**
   *
   * @type {string}
   * @memberof PaginatedLearningResourceList
   */
  previous?: string | null
  /**
   *
   * @type {Array<LearningResource>}
   * @memberof PaginatedLearningResourceList
   */
  results: Array<LearningResource>
}

/**
 * Check if a given object implements the PaginatedLearningResourceList interface.
 */
export function instanceOfPaginatedLearningResourceList(
  value: object,
): value is PaginatedLearningResourceList {
  if (!("count" in value) || value["count"] === undefined) return false
  if (!("results" in value) || value["results"] === undefined) return false
  return true
}

export function PaginatedLearningResourceListFromJSON(
  json: any,
): PaginatedLearningResourceList {
  return PaginatedLearningResourceListFromJSONTyped(json, false)
}

export function PaginatedLearningResourceListFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PaginatedLearningResourceList {
  if (json == null) {
    return json
  }
  return {
    count: json["count"],
    next: json["next"] == null ? undefined : json["next"],
    previous: json["previous"] == null ? undefined : json["previous"],
    results: (json["results"] as Array<any>).map(LearningResourceFromJSON),
  }
}

export function PaginatedLearningResourceListToJSON(
  json: any,
): PaginatedLearningResourceList {
  return PaginatedLearningResourceListToJSONTyped(json, false)
}

export function PaginatedLearningResourceListToJSONTyped(
  value?: PaginatedLearningResourceList | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    count: value["count"],
    next: value["next"],
    previous: value["previous"],
    results: (value["results"] as Array<any>).map(LearningResourceToJSON),
  }
}